move dynamo tests to seperate files according to function
